% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cateditmatrix.R
\name{parseCatEdit}
\alias{parseCatEdit}
\title{parse categorial edit}
\usage{
parseCatEdit(el)
}
\arguments{
\item{el}{\code{expression} with a single edit}
}
\value{
named \code{numeric} with coefficients
}
\description{
parse categorial edit
}
\keyword{internal}
