% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{[.editmatrix}
\alias{[.editmatrix}
\alias{[.cateditmatrix}
\alias{[.editarray}
\alias{[.editset}
\alias{[.editlist}
\title{Row index operator for \code{editmatrix}}
\usage{
\method{[}{editmatrix}(x, i, j, ...)

\method{[}{editarray}(x, i, j, ...)

\method{[}{editarray}(x, i, j, ...)

\method{[}{editset}(x, i, j, ...)

\method{[}{editlist}(x, i, j, ...)
}
\arguments{
\item{x}{an object of class \code{\link{editmatrix}} or \code{\link{editarray}}}

\item{i}{the row index in the edit matrix (numeric, logical or rowname)}

\item{j}{the column index in the edit matrix}

\item{...}{arguments to be passed to other methods. Currently ignored.}
}
\description{
Use this operator to select edits from an editmatrix or editarray object.
}
\keyword{internal}
